Add files via upload
I made this to keep track of an achievement in a game. I had to keep track of a list of areas using a different combination of letters 50 times (It makes more sense if you're familiar with the game. Game title is .hack//G.U. Last Recode and the achievement is called Clear All Areas). I wrote it in 2 hours.