Add files via upload
This is a program that compares two lists of IP addresses and prints out a text file with what it found. It's functional, but it needs a user interface and another user input check. It uses text files that are in the same directory. After getting the interface working, I'll plan on adjusting it to where it can pull online WordLists to compare instead of text files. I'm also having some issues getting JavaFX to work in JDK 15. I think I started this project with JDK 10, so any help/articles would be appreciated.